pyproject.toml
[project]
name = "yaap"
version = "0.1.0"
description = "AI powered penetration testing assistant"
authors = [ { name = "findthehead" } ]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "fpdf2>=2.7.9",
  "langchain>=0.1.0",
  "langchain-anthropic>=1.3.5",
  "langchain-core>=0.1.0",
  "langchain-google-genai>=0.0.1",
  "langchain-ollama>=0.0.1",
  "langchain-openai>=0.0.1",
  "langgraph>=0.0.1",
  "markdown>=3.5.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.0",
  "pytest>=7.0.0",
  "requests>=2.28.0",
]

[tool.uv]
# Lock file will be created as uv.lock after first sync

[tool.uv.sources]
# Example for private indexes if needed:
# pypi = { url = "https://pypi.org/simple" }

[tool.uv.workspace]
members = []
